Parameter Configuration of Mosmafs

Create a SMOOF function for parameter configuration of mosmafs, with parameter set.

The resulting function takes a list of values according to its getParamSet(). Additionally the list can contain an $INSTANCE, an integer between 1 and 1000. If it is not given, the instance will be chosen randomly. It corresponds to the resampling instance to use if fixed.ri is TRUE.

constructEvalSetting(
  task,
  learner,
  ps,
  measure = getDefaultMeasure(task),
  worst.measure = NULL,
  cpo = NULLCPO,
  nfeat = getTaskNFeats(task %>>% cpo),
  evals = 1e+05,
  outer.resampling = makeResampleDesc("CV", iters = 10, stratify = TRUE),
  savedir = NULL
)

Arguments

task
[Task] the task to optimize.

learner
[Learner] the learner to optimize.

ps
[ParamSet] the parameter set of the learner (and cpo) alone.

measure
[Measure] measure to optimize.

worst.measure
[numeric(1)] worst value for measure to consider, for dominated hypervolume calculation. Will be extracted from the given measure if not given, but will raise an error if the extracted (or given) value is infinite.

cpo
[CPO] cpo to prepend feature selection.

nfeat
[integer(1)] number of features.

evals
[integer(1)] number of evals to perform. Note this concerns fidelity evaluations (i.e. single CV folds). When not using multifid the number of points evaluated is 1/10th the evals value.

outer.resampling
outer resampling to use.

savedir
[character(1) | NULL] the directory to save every trace to. If this is NULL (the default) evaluations are not saved.

Value

function a smoof function.